Initialized Remix app using "Indie" stack
THE INDIE STACK
Deployed to a long-running Node.js server with a persistent SQLite database.
This stack is great for websites with dynamic data that you control (blogs, marketing, content sites).
It's also a perfect, low-complexity bootstrap for MVPs, prototypes, and proof-of-concepts that can later be updated to the Blues stack easily.